Tempory commit, just to contain and push up my current workings out.
